% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imperial_colour_scheme_resources.R
\name{imperial_cols}
\alias{imperial_cols}
\title{Function to extract imperial colors as hex codes}
\usage{
imperial_cols(...)
}
\arguments{
\item{...}{Character names of imperial_colors}
}
\description{
Function to extract imperial colors as hex codes
}
