% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imperial_colour_scheme_functions.R
\name{imperial_pal}
\alias{imperial_pal}
\title{Return function to interpolate a new color palette}
\usage{
imperial_pal(palette = "default", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in imperial_palettes. Choose from:
\itemize{
 \item default, earth, cool, warm, soft, bold, nature,
 \item two_col_grey_teal, two_col_pink_purple, two_col_blue_green
}}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a new color palette
}
