\name{legacy_helpers}
\alias{GAMModelMaker}
\alias{ModelMakerMulti}
\alias{ModelMakerMultiRD}
\alias{add_conf_ints}
\alias{binaryCleaner}
\alias{binaryCleaner_1_0}
\alias{build_cumulative_sets}
\alias{calculateEpiStats}
\alias{calculate_prop_ci}
\alias{canBeNumeric}
\alias{checkConvertNumeric}
\alias{check_perfect_separation}
\alias{comma_sep}
\alias{compute_level_order}
\alias{continuousCleaner}
\alias{crossTab}
\alias{crossTabContinuous}
\alias{crossTabMulti}
\alias{dateCleaner}
\alias{get_pretty_name}
\alias{imperial_colours}
\alias{imperial_cols}
\alias{imperial_palettes}
\alias{makeORTable}
\alias{makeRDTable}
\alias{makeStratifiedPrevalenceTablePlottable}
\alias{makeXtabPlottable}
\alias{modelMakerAdjustedSetsRD}
\alias{modelMakerSequential}
\alias{modelMakerSequentialRD}
\alias{naReporter}
\alias{orderMyTable}
\alias{pValEpiConverter}
\alias{plotReactForest}
\alias{plotReactVIF}
\alias{pvalAsterisker}
\alias{reactCleaner}
\alias{run_auto_tableone_regressions}
\alias{safe_makeRDTable}
\alias{savePrettyExcelWorkbook}
\alias{saveREACT}
\alias{saveREACTplot}
\alias{saveREACTtable}
\alias{scale_color_imperial}
\alias{scale_fill_imperial}
\alias{stratifiedTables}
\alias{theme_mw}
\alias{theme_react}
\alias{validate_factor_levels}
\alias{xtabPercentageExtractor}
\title{Legacy and Utility Helpers}
\description{
This topic groups exported helpers and objects retained for compatibility
across the package.
}
\details{
The aliases on this page cover legacy cleaning, tabulation, modelling,
plotting, palette, and export helpers that are still exported by the package.
See the corresponding source files in the \code{R/} directory for implementation
details.
}
