% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prevalence_tables.R
\name{makeTablesNew}
\alias{makeTablesNew}
\title{Function created on 23-2-2021 by Matt Whitaker, to work with new wrangling system for REACT-2
cov_name_list can be found in cats_and_covs.R and is a named list of all useful covariates in react2
Any named list of covariates with variable descriptions will work here, or the function will work without}
\usage{
makeTablesNew(
  dat = dfRes,
  result_var = "res",
  pos_val = "1",
  neg_val = "0",
  covariates = covs,
  cov_name_list,
  sens = 0.844,
  spec = 0.986,
  weights = NULL,
  suffix = NULL,
  sf = 4,
  for_report = FALSE,
  write_to_file = FALSE,
  percent = TRUE,
  output_list = TRUE,
  separator = ","
)
}
\description{
Relies on Kylie's add_conf_ints function
}
\details{
Make prevalence tables
for instances where there are only zeros or ones in the results (ie all pos or all neg)
}
