% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_OR_tables.R
\name{ModelMakerMulti}
\alias{ModelMakerMulti}
\title{wrapper function for the modelMakerSequential function, which allows you to pass a list of
variables of interest and return 1) a data frame of ORs and concatenated CIs for each level of variable of interest,
2) a big df of ORs with separate CIs for plotting as a forest plot}
\usage{
ModelMakerMulti(
  dat = dfRes,
  list_of_variables_of_interest,
  outcome = "res",
  joint_adjustment_vars = c("age_group_named", "sex", "region_named", "ethnic_new",
    "imd_quintile_cat"),
  cov_name_list = NULL
)
}
\description{
wrapper function for the modelMakerSequential function, which allows you to pass a list of
variables of interest and return 1) a data frame of ORs and concatenated CIs for each level of variable of interest,
2) a big df of ORs with separate CIs for plotting as a forest plot
}
