% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imperial_colour_scheme_functions.R
\name{scale_color_imperial}
\alias{scale_color_imperial}
\title{Color scale constructor for imperial colors}
\usage{
scale_color_imperial(
  palette = "default",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in imperial_palettes. Choose from:
\itemize{
 \item default, core, cool, warm, extended,
 \item pink, green, red, blue,
 \item tbp, vbn, obib, bnt,
 \item bgrey_grad, bgrn_grad
}}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to ggplot2::discrete_scale() or
ggplot2::scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for imperial colors
}
