% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imperial_colour_scheme_functions.R
\name{scale_color_new}
\alias{scale_color_new}
\title{Color scale constructor for new colors}
\usage{
scale_color_new(palette = "default", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in imperial_palettes. Choose from:
\itemize{
 \item default, earth, cool, warm, soft, bold, nature,
 \item two_col_grey_teal, two_col_pink_purple, two_col_blue_green
}}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to ggplot2::discrete_scale() or
ggplot2::scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for new colors
}
