% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prevalence_tables.R, R/specify_decimal.R
\name{specifyDecimal}
\alias{specifyDecimal}
\title{Function created on 23-2-2021 by Matt Whitaker, to work with new wrangling system for REACT-2
cov_name_list can be found in cats_and_covs.R and is a named list of all useful covariates in react2
Any named list of covariates with variable descriptions will work here, or the function will work without}
\usage{
specifyDecimal(x, k)

specifyDecimal(x, k)
}
\description{
Relies on Kylie's add_conf_ints function
}
