% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prevalence_tables.R
\name{makeTablesNew}
\alias{makeTablesNew}
\title{Make prevalence tables
for instances where there are only zeros or ones in the results (ie all pos or all neg)}
\usage{
makeTablesNew(
  dat = dfRes,
  result_var = "res",
  pos_val = "1",
  neg_val = "0",
  covariates = covs,
  cov_name_list,
  sens = 0.844,
  spec = 0.986,
  weights = NULL,
  suffix = NULL,
  sf = 4,
  for_report = FALSE,
  write_to_file = FALSE,
  percent = TRUE,
  output_list = TRUE
)
}
\description{
Make prevalence tables
for instances where there are only zeros or ones in the results (ie all pos or all neg)
}
