% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_decimal.R
\name{specifyDecimal}
\alias{specifyDecimal}
\title{Function created on 23-2-2021 by Matt Whitaker, to work with new wrangling system for REACT-2}
\usage{
specifyDecimal(x, k, format = "fg", simpleround = F)
}
\description{
Function created on 23-2-2021 by Matt Whitaker, to work with new wrangling system for REACT-2
}
